+++ packages/db/src/public/PublicSchema.ts
@@ -33,22 +33,6 @@ import type { StagesTable } from "./Stages";
 import type { UsersTable } from "./Users";
 
 export interface PublicSchema {
-	_prisma_migrations: PrismaMigrationsTable;
-
-	users: UsersTable;
-
-	pubs: PubsTable;
-
-	pub_types: PubTypesTable;
-
-	stages: StagesTable;
-
-	member_groups: MemberGroupsTable;
-
-	communities: CommunitiesTable;
-
-	move_constraint: MoveConstraintTable;
-
 	pub_fields: PubFieldsTable;
 
 	pub_values: PubValuesTable;
@@ -92,4 +76,20 @@ export interface PublicSchema {
 	membership_capabilities: MembershipCapabilitiesTable;
 
 	pub_values_history: PubValuesHistoryTable;
+
+	_prisma_migrations: PrismaMigrationsTable;
+
+	users: UsersTable;
+
+	pubs: PubsTable;
+
+	pub_types: PubTypesTable;
+
+	stages: StagesTable;
+
+	member_groups: MemberGroupsTable;
+
+	communities: CommunitiesTable;
+
+	move_constraint: MoveConstraintTable;
 }
